# Lab4/ark-pzpi-23-5-butkevych-anna-lab4/RadiationMonitoring/api-docs-latest.json
{"openapi":"3.0.0","info":{"title":"Radiation Monitoring API","version":"1.0.0","description":"API for radiation monitoring system with real-time alerts"},"servers":[{"url":"http://localhost:3000/api","description":"Development server"}],"tags":[{"name":"Auth","description":"Authentication endpoints (register, login)"},{"name":"Users","description":"User management endpoints"},{"name":"Devices","description":"Device management endpoints"},{"name":"Readings","description":"Sensor readings endpoints"},{"name":"Alerts","description":"Alerts and notifications"},{"name":"Subscriptions","description":"User subscription endpoints"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"},"deviceToken":{"type":"apiKey","in":"header","name":"Device-Token"}},"schemas":{"User":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string","format":"email"},"name":{"type":"string"},"role":{"type":"string"}}},"Device":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"device_token":{"type":"string"},"is_active":{"type":"boolean"}}},"SensorReading":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"device_id":{"type":"string","format":"uuid"},"value":{"type":"number"},"unit":{"type":"string"},"measured_at":{"type":"string","format":"date-time"}}},"Alert":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"device_id":{"type":"string","format":"uuid"},"level":{"type":"string"},"message":{"type":"string"},"acknowledged":{"type":"boolean"}}}}},"paths":{"/api/alerts":{"get":{"tags":["Alerts"],"summary":"Get alerts for user's devices","security":[{"bearerAuth":[]}],"parameters":[{"in":"query","name":"level","schema":{"type":"string","enum":["warning","danger","critical"]},"description":"Filter by alert level"},{"in":"query","name":"acknowledged","schema":{"type":"boolean"},"description":"Filter by acknowledgment status"},{"in":"query","name":"limit","schema":{"type":"integer","default":50},"description":"Number of alerts to return"}],"responses":{"200":{"description":"List of alerts","content":{"application/json":{"examples":{"success":{"summary":"Alerts list","value":{"alerts":[{"id":"uuid","level":"warning","message":"High reading"}]}}}}}}}}},"/api/alerts/{id}":{"get":{"tags":["Alerts"],"summary":"Get alert by ID","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Alert details"},"404":{"description":"Alert not found"}}}},"/api/alerts/{id}/ack":{"post":{"tags":["Alerts"],"summary":"Acknowledge an alert","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Alert acknowledged successfully"},"404":{"description":"Alert not found"}}}},"/api/auth/register":{"post":{"tags":["Auth"],"summary":"Register a new user","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":6},"name":{"type":"string"},"role":{"type":"string","enum":["user","admin"]}}},"examples":{"sample":{"summary":"Sample register payload","value":{"email":"user@example.com","password":"password123","name":"Example User"}}}}}},"responses":{"201":{"description":"User registered successfully","content":{"application/json":{"examples":{"success":{"summary":"Successful response","value":{"message":"User registered successfully","user":{"id":"uuid","email":"user@example.com","name":"Example User"},"token":"<jwt>"}}}}}},"400":{"description":"Validation error"},"409":{"description":"Email already exists"}}}},"/api/auth/login":{"post":{"tags":["Auth"],"summary":"Login user","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string"}}},"examples":{"sample":{"summary":"Sample login payload","value":{"email":"user@example.com","password":"password123"}}}}}},"responses":{"200":{"description":"Login successful","content":{"application/json":{"examples":{"success":{"summary":"Successful login","value":{"message":"Login successful","user":{"id":"uuid","email":"user@example.com"},"token":"<jwt>"}}}}}},"400":{"description":"Invalid credentials"}}}},"/api/devices":{"get":{"tags":["Devices"],"summary":"Get all devices for authenticated user","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of devices","content":{"application/json":{"examples":{"success":{"summary":"List of devices","value":{"devices":[{"id":"uuid","name":"Device 1","device_token":"token","is_active":true}]}}}}}}}},"post":{"tags":["Devices"],"summary":"Create a new device","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"location_id":{"type":"string","format":"uuid"}}},"examples":{"sample":{"summary":"Create device payload","value":{"name":"Device 1","location_id":"uuid"}}}}}},"responses":{"201":{"description":"Device created successfully","content":{"application/json":{"examples":{"success":{"summary":"Created device","value":{"message":"Device created successfully","device":{"id":"uuid","name":"Device 1","device_token":"token"}}}}}}}}}},"/api/devices/{id}":{"get":{"tags":["Devices"],"summary":"Get device by ID","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Device details"},"404":{"description":"Device not found"}}},"put":{"tags":["Devices"],"summary":"Update device","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"location_id":{"type":"string","format":"uuid"},"is_active":{"type":"boolean"}}}}}},"responses":{"200":{"description":"Device updated successfully"},"404":{"description":"Device not found"}}},"delete":{"tags":["Devices"],"summary":"Delete device","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Device deleted successfully"},"404":{"description":"Device not found"}}}},"/api/readings":{"post":{"tags":["Readings"],"summary":"Submit sensor reading from device","security":[{"deviceToken":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"measured_at":{"type":"string","format":"date-time"},"value":{"type":"number"},"unit":{"type":"string"},"metadata":{"type":"object"}}},"examples":{"sample":{"summary":"Sample reading","value":{"measured_at":"2025-12-16T12:00:00Z","value":0.8,"unit":"µSv/h"}}}}}},"responses":{"201":{"description":"Reading recorded successfully","content":{"application/json":{"examples":{"success":{"summary":"Created reading","value":{"message":"Reading recorded successfully","reading":{"id":"uuid","device_id":"uuid","value":0.8}}}}}}}}},"get":{"tags":["Readings"],"summary":"Get readings for user's devices","security":[{"bearerAuth":[]}],"parameters":[{"in":"query","name":"device_id","schema":{"type":"string"},"description":"Filter by device ID"},{"in":"query","name":"limit","schema":{"type":"integer","default":100},"description":"Number of readings to return"},{"in":"query","name":"since","schema":{"type":"string","format":"date-time"},"description":"Get readings since this date"}],"responses":{"200":{"description":"List of readings","content":{"application/json":{"examples":{"success":{"summary":"Readings list","value":{"readings":[{"id":"uuid","device_id":"uuid","value":0.8}]}}}}}}}}},"/api/subscriptions":{"post":{"tags":["Subscriptions"],"summary":"Create a new subscription","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"channel":{"type":"string","enum":["email","sms","push"]},"criteria":{"type":"object","properties":{"levels":{"type":"array","items":{"type":"string","enum":["warning","danger","critical"]}},"threshold":{"type":"number","description":"Custom threshold value in µSv/h"}}},"active":{"type":"boolean","default":true}}},"examples":{"sample":{"summary":"Subscription payload","value":{"channel":"email","criteria":{"levels":["warning","danger"]}}}}}}},"responses":{"201":{"description":"Subscription created successfully"}}},"get":{"tags":["Subscriptions"],"summary":"Get user's subscriptions","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of subscriptions","content":{"application/json":{"examples":{"success":{"summary":"Subscriptions list","value":{"subscriptions":[{"id":"uuid","channel":"email"}]}}}}}}}}},"/api/subscriptions/{id}":{"delete":{"tags":["Subscriptions"],"summary":"Delete a subscription","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Subscription deleted successfully"},"404":{"description":"Subscription not found"}}},"put":{"tags":["Subscriptions"],"summary":"Update a subscription","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"channel":{"type":"string","enum":["email","sms","push"]},"criteria":{"type":"object"},"active":{"type":"boolean"}}}}}},"responses":{"200":{"description":"Subscription updated successfully"},"404":{"description":"Subscription not found"}}}}}}